Rename YARG.Core.Song.Deserialization to YARG.Core.IO
