+++ YARG.Core/IO/YARGCONLoader.cs
@@ -4,7 +4,7 @@ using System.Diagnostics;
 using System.IO;
 using System.Text;
 
-namespace YARG.Core.Song.Deserialization
+namespace YARG.Core.IO
 {
     public enum CONFileListingFlag : byte
     {
