+++ YARG.Core/IO/YARGChartFileReader.cs
@@ -2,9 +2,9 @@
 using System.Collections.Generic;
 using System.Linq;
 using System.Text;
-using YARG.Core.Song.Deserialization.Ini;
+using YARG.Core.IO.Ini;
 
-namespace YARG.Core.Song.Deserialization
+namespace YARG.Core.IO
 {
     public enum ChartEventType
     {
