+++ YARG.Core/IO/YARGMidiReader.cs
@@ -6,7 +6,7 @@ using System.IO;
 using System.Text;
 using YARG.Core.Extensions;
 
-namespace YARG.Core.Song.Deserialization
+namespace YARG.Core.IO
 {
     public enum MidiEventType : byte
     {
