+++ YARG.Core/Song/Cache/CacheNodes.cs
@@ -2,7 +2,7 @@
 using System.IO;
 using System.Threading.Tasks;
 using YARG.Core.Extensions;
-using YARG.Core.Song.Deserialization;
+using YARG.Core.IO;
 
 namespace YARG.Core.Song.Cache
 {
