+++ YARG.Core/Song/Metadata/AvailableParts/AvailableParts.Chart.cs
@@ -1,6 +1,6 @@
 ﻿using System;
 using YARG.Core.Chart;
-using YARG.Core.Song.Deserialization;
+using YARG.Core.IO;
 using YARG.Core.Song.Preparsers;
 
 namespace YARG.Core.Song
