+++ YARG.Core/Song/Metadata/AvailableParts/AvailableParts.RBCON.cs
@@ -1,4 +1,4 @@
-﻿using YARG.Core.Song.Deserialization;
+﻿using YARG.Core.IO;
 
 namespace YARG.Core.Song
 {
