+++ YARG.Core/Song/Preparsers/DrumPreparseHandler.cs
@@ -1,6 +1,6 @@
 ﻿using System;
 using YARG.Core.Chart;
-using YARG.Core.Song.Deserialization;
+using YARG.Core.IO;
 
 namespace YARG.Core.Song.Preparsers
 {
