+++ YARG.Core/Song/Preparsers/Midi/MidiFiveFretPreparser.cs
@@ -1,6 +1,6 @@
 ﻿using System;
 using System.Text;
-using YARG.Core.Song.Deserialization;
+using YARG.Core.IO;
 
 namespace YARG.Core.Song
 {
