+++ YARG.Core/Song/Preparsers/Midi/MidiVocalPreparser.cs
@@ -1,5 +1,5 @@
 ﻿using System;
-using YARG.Core.Song.Deserialization;
+using YARG.Core.IO;
 
 namespace YARG.Core.Song
 {
